+++ app/src/main/java/com/example/servicesnovigrad/WeeklySchedule.java
@@ -8,18 +8,18 @@ import java.util.ArrayList;
 import java.util.EnumMap;
 
 public class WeeklySchedule {
-    private EnumMap<DayOfWeek, ArrayList<Pair<Short, Short>>> openHours;
+    private EnumMap<DayOfWeek, ArrayList<Pair<Integer, Integer>>> openHours;
 
     public WeeklySchedule() {
-        openHours = new EnumMap<DayOfWeek, ArrayList<Pair<Short, Short>>>(DayOfWeek.class);
+        openHours = new EnumMap<DayOfWeek, ArrayList<Pair<Integer, Integer>>>(DayOfWeek.class);
     }
 
-    public ArrayList<Pair<Short, Short>> getOpenHours(DayOfWeek day) {
+    public ArrayList<Pair<Integer, Integer>> getOpenHours(DayOfWeek day) {
         if (!openHours.containsKey(day))
-            openHours.put(day, new ArrayList<Pair<Short, Short>>());
+            openHours.put(day, new ArrayList<Pair<Integer, Integer>>());
         return openHours.get(day);
     }
-    public EnumMap<DayOfWeek, ArrayList<Pair<Short, Short>>> getOpenHours() {
+    public EnumMap<DayOfWeek, ArrayList<Pair<Integer, Integer>>> getOpenHours() {
         return openHours;
     }
 
@@ -27,21 +27,21 @@ public class WeeklySchedule {
         Adds a pair of openingTime, closingTime for a specific day of the week;
         Time is in minutes from 0:00
     */
-    public void addOpenHours(DayOfWeek day, short openingTime, short closingTime) {
+    public void addOpenHours(DayOfWeek day, int openingTime, int closingTime) {
         if (openingTime > closingTime)
             throw new InvalidParameterException("ERROR : Opening time is larger than closing time");
         if (openingTime >= 60 * 24) openingTime = 60 * 24 - 1;
         if (closingTime >= 60 * 24) closingTime = 60 * 24 - 1;
 
         if (!openHours.containsKey(day))
-            openHours.put(day, new ArrayList<Pair<Short, Short>>());
-        openHours.get(day).add(new Pair<Short, Short>(openingTime, closingTime));
+            openHours.put(day, new ArrayList<Pair<Integer, Integer>>());
+        openHours.get(day).add(new Pair<Integer, Integer>(openingTime, closingTime));
     }
     /*
         Removes the pair of openingTime, closingTime of index
         in the array contained for a particular day
     */
-    public Pair<Short, Short> removeOpenHours(DayOfWeek day, int index){
+    public Pair<Integer, Integer> removeOpenHours(DayOfWeek day, int index){
         if(index<0)
             throw new NullPointerException("Cannot access this position: " + index);
         if (!openHours.containsKey(day) || openHours.get(day).size()<=index)
@@ -57,7 +57,7 @@ public class WeeklySchedule {
              DayOfWeek.values()) {
             s += day + "/n";
             if (openHours.containsKey(day))
-            for (Pair<Short, Short> p:
+            for (Pair<Integer, Integer> p:
             openHours.get(day)) {
                 s += " from " + p.first/60 + ":" + p.first%60;
                 s += " to " + p.second/60 + ":" + p.second%60 + "/n";
